src/requirements.txt
aiohttp==3.7.3
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.4.7
decorator==4.4.2
discord.py==1.6.0
distlib==0.3.1
enum34==1.1.10
filelock==3.0.12
flake8==3.8.4
future==0.18.2
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.6.1
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.2
parso==0.8.1
pathspec==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
plyer==2.0.0
prompt-toolkit==3.0.14
psutil==5.8.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pylint==2.6.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.19.0
pytube==11.0.0
pytz==2020.5
pyzmq==22.0.2
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
tornado==6.1
traitlets==5.0.5
tweepy==3.10.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.0
wcwidth==0.2.5
wrapt==1.12.1
yapf==0.30.0
yarl==1.6.3
